Restructuring Transpose Conv integration files. Registering Int8, Int16 data-types seperately and integrating & registering Float32 data-type. Also registering reference implementations of all 3 mentioned data-types.
